refactor(component): remove barrels from exporting to the public API
Closes #2643
