--- modules/component/src/core/cd-aware/index.ts
@@ -1,2 +1,0 @@
-export * from './creator_render';
-export * from './cd-aware_creator';
