--- modules/component/src/core/index.ts
@@ -1,2 +1,0 @@
-export * from './utils';
-export * from './cd-aware';
