--- modules/component/src/core/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './has-zone';
