+++ modules/component/src/index.ts
@@ -1,3 +1,3 @@
-export { PushPipe } from './push';
-export { LetDirective } from './let';
+export { PushPipe } from './push/push.pipe';
+export { LetDirective } from './let/let.directive';
 export { ReactiveComponentModule } from './reactive-component.module';
