--- modules/component/src/let/index.ts
@@ -1,1 +1,0 @@
-export * from './let.directive';
