--- modules/component/src/push/index.ts
@@ -1,1 +1,0 @@
-export * from './push.pipe';
