+++ modules/component/src/reactive-component.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
-import { LetDirective } from './let';
-import { PushPipe } from './push';
+
+import { LetDirective } from './let/let.directive';
+import { PushPipe } from './push/push.pipe';
 
 const DECLARATIONS = [LetDirective, PushPipe];
 const EXPORTS = [DECLARATIONS];
